refactor(shell): extract SidePanel from App shell
